+++ core/src/domain/entities/message/mod.rs
@@ -1,3 +1,3 @@
 mod message_sending;
 
-pub use message_sending::MessageSending;
+pub use message_sending::*;