+++ core/src/use_cases/role_scoped/guest_manager/token/create_role_associated_connection_string.rs
@@ -8,7 +8,7 @@ use crate::{
             route_type::PermissionedRoles,
             token::{RoleScopedConnectionString, RoleWithPermissionsScope},
         },
-        entities::{MessageSending, TokenRegistration},
+        entities::{LocalMessageSending, TokenRegistration},
     },
     models::AccountLifeCycle,
     use_cases::support::send_email_notification,
@@ -34,7 +34,7 @@ pub async fn create_role_associated_connection_string(
     permissioned_roles: PermissionedRoles,
     life_cycle_settings: AccountLifeCycle,
     token_registration_repo: Box<&dyn TokenRegistration>,
-    message_sending_repo: Box<&dyn MessageSending>,
+    message_sending_repo: Box<&dyn LocalMessageSending>,
 ) -> Result<String, MappedErrors> {
     // ? -----------------------------------------------------------------------
     // ? Check if the current account has sufficient privileges to create role