+++ ports/api/src/main.rs
@@ -38,10 +38,16 @@ use endpoints::{
     staff::account_endpoints as staff_account_endpoints,
 };
 use models::config_handler::ConfigHandler;
+use myc_adapters_shared_lib::models::{
+    SharedAppModule, SharedClientImpl, SharedClientImplParameters,
+    SharedClientProvider,
+};
 use myc_config::{
     init_vault_config_from_file, optional_config::OptionalConfig,
 };
-use myc_core::settings::init_in_memory_routes;
+use myc_core::{
+    domain::entities::RemoteMessageSending, settings::init_in_memory_routes,
+};
 use myc_diesel::repositories::{
     DieselDbPoolProvider, DieselDbPoolProviderParameters, SqlAppModule,
 };
@@ -49,8 +55,12 @@ use myc_http_tools::{
     providers::{azure_endpoints, google_endpoints},
     settings::DEFAULT_REQUEST_ID_KEY,
 };
-use myc_notifier::settings::{
-    init_queue_config_from_file, init_smtp_config_from_file,
+use myc_kv::repositories::KVAppModule;
+use myc_notifier::{
+    models::ClientProvider,
+    repositories::{
+        NotifierAppModule, NotifierClientImpl, NotifierClientImplParameters,
+    },
 };
 use oauth2::http::HeaderName;
 use openssl::{
@@ -66,6 +76,7 @@ use reqwest::header::{
 };
 use router::route_request;
 use settings::{ADMIN_API_SCOPE, GATEWAY_API_SCOPE, SUPER_USER_API_SCOPE};
+use shaku::HasComponent;
 use std::{path::PathBuf, str::FromStr, sync::Arc};
 use tracing::{info, trace};
 use tracing_actix_web::TracingLogger;
@@ -165,16 +176,17 @@ pub async fn main() -> std::io::Result<()> {
     //
     // ? -----------------------------------------------------------------------
     info!("Initializing SMTP configs");
-    init_smtp_config_from_file(None, Some(config.smtp)).await;
+    //init_smtp_config_from_file(None, Some(config.smtp)).await;
 
     info!("Initializing QUEUE configs");
-    init_queue_config_from_file(None, Some(config.queue.to_owned())).await;
+    //init_queue_config_from_file(None, Some(config.queue.to_owned())).await;
 
     // ? -----------------------------------------------------------------------
     // ? Configure SQL App Module
     // ? -----------------------------------------------------------------------
 
-    info!("Initialize SQL dependencies");
+    info!("Initialize internal dependencies");
+
     let sql_module = Arc::new(
         SqlAppModule::builder()
             .with_component_parameters::<DieselDbPoolProvider>(
@@ -198,6 +210,62 @@ pub async fn main() -> std::io::Result<()> {
             .build(),
     );
 
+    let shared_provider =
+        match SharedClientImpl::new(config.redis.to_owned()).await {
+            Ok(provider) => provider,
+            Err(err) => panic!("Error on initialize shared provider: {err}"),
+        };
+
+    let shared_module = Arc::new(
+        SharedAppModule::builder()
+            .with_component_parameters::<SharedClientImpl>(
+                SharedClientImplParameters {
+                    redis_client: shared_provider.get_redis_client(),
+                    redis_config: shared_provider.get_redis_config(),
+                },
+            )
+            .build(),
+    );
+
+    let notifier_provider = match NotifierClientImpl::new(
+        config.queue.to_owned(),
+        config.redis.to_owned(),
+        config.smtp.to_owned(),
+    )
+    .await
+    {
+        Ok(provider) => provider,
+        Err(err) => panic!("Error on initialize notifier provider: {err}"),
+    };
+
+    let notifier_module = Arc::new(
+        NotifierAppModule::builder()
+            .with_component_parameters::<SharedClientImpl>(
+                SharedClientImplParameters {
+                    redis_client: shared_provider.get_redis_client(),
+                    redis_config: shared_provider.get_redis_config(),
+                },
+            )
+            .with_component_parameters::<NotifierClientImpl>(
+                NotifierClientImplParameters {
+                    smtp_client: notifier_provider.get_smtp_client(),
+                    queue_config: notifier_provider.get_queue_config(),
+                },
+            )
+            .build(),
+    );
+
+    let kv_module = Arc::new(
+        KVAppModule::builder()
+            .with_component_parameters::<SharedClientImpl>(
+                SharedClientImplParameters {
+                    redis_client: shared_provider.get_redis_client(),
+                    redis_config: shared_provider.get_redis_config(),
+                },
+            )
+            .build(),
+    );
+
     // ? -----------------------------------------------------------------------
     // ? Fire the scheduler
     //
@@ -206,7 +274,20 @@ pub async fn main() -> std::io::Result<()> {
     //
     // ? -----------------------------------------------------------------------
     info!("Fire email dispatcher");
-    email_dispatcher(config.queue.to_owned());
+
+    email_dispatcher(
+        config.queue.to_owned(),
+        unsafe {
+            Arc::from_raw(*Arc::new(
+                notifier_module.resolve_ref() as &dyn SharedClientProvider
+            ))
+        },
+        unsafe {
+            Arc::from_raw(*Arc::new(
+                notifier_module.resolve_ref() as &dyn RemoteMessageSending
+            ))
+        },
+    );
 
     // ? -----------------------------------------------------------------------
     // ? Fire the scheduler
@@ -265,6 +346,9 @@ pub async fn main() -> std::io::Result<()> {
             //
             .wrap(TracingLogger::default())
             .app_data(web::Data::from(sql_module.clone()))
+            .app_data(web::Data::from(shared_module.clone()))
+            .app_data(web::Data::from(notifier_module.clone()))
+            .app_data(web::Data::from(kv_module.clone()))
             .app_data(web::Data::new(token_config).clone())
             .app_data(web::Data::new(auth_config.to_owned()).clone())
             //