+++ ports/api/src/modules/message.rs
@@ -1,9 +1,0 @@
-use myc_notifier::repositories::MessageSendingQueueRepository;
-use shaku::module;
-
-module! {
-    pub MessageSendingQueueModule {
-        components = [MessageSendingQueueRepository],
-        providers = []
-    }
-}