+++ ports/api/src/modules/mod.rs
@@ -1,5 +1,3 @@
-mod message;
 mod route;
 
-pub use message::*;
 pub use route::*;